+++ src/pythonwrench/argparse.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+import re
 from argparse import ArgumentParser
 from dataclasses import MISSING, fields
 from functools import partial
@@ -25,13 +26,15 @@ except ImportError:
     # support older python versions
     UnionType = Any
 
+from pythonwrench.functools import filter_and_call
 from pythonwrench.typing.classes import Dataclass, DataclassInstance, NoneType
+from pythonwrench.warnings import deprecated_alias
 
 T = TypeVar("T")
 T_Dataclass = TypeVar("T_Dataclass", bound=Dataclass)
 T_DataclassInstance = TypeVar("T_DataclassInstance", bound=DataclassInstance)
-TargetType = Union[Type[T], UnionType, "Type[Literal]"]
-
+TargetType = Union[Type[T], UnionType, "Type[Literal]", "Type[Optional]"]
+ListParsing = Literal["argparse", "brackets"]
 
 DEFAULT_TRUE_VALUES = ("True", "t", "yes", "y", "1")
 DEFAULT_FALSE_VALUES = ("False", "f", "no", "n", "0")
@@ -42,24 +45,41 @@ _SCALARS_TARGET_TYPES = (str, int, float, None, NoneType, bool)
 
 def parse_args_using_dataclass(
     dataclass_type: Type[T_DataclassInstance],
+    *,
     args: Optional[Iterable[str]] = None,
+    parser: Optional[ArgumentParser] = None,
+    list_parsing: ListParsing = "argparse",
 ) -> T_DataclassInstance:
     """Converts prog args to a typed dataclass using argparse.
 
     Currently only supports dataclasses that contains only builtin scalars: str, int, float, None, bool OR list of builtin scalars.
     """
-    parser = ArgumentParser()
-    parser = new_parser_from_dataclass(dataclass_type, parser)
+    init_parser = parser
+    parser = add_dataclass_fields_to_parser(
+        dataclass_type,
+        parser=parser,
+        list_parsing=list_parsing,
+    )
     parsed, argv = parser.parse_known_args(args)
     if len(argv) > 0:
         raise ValueError(f"Found {len(argv)} unknown arguments: {argv}.")
-    instance = dataclass_type(**parsed.__dict__)
+
+    if init_parser is None:
+        instance = dataclass_type(**parsed.__dict__)
+    else:
+        instance = filter_and_call(
+            dataclass_type,
+            _fill_all_arguments=True,
+            **parsed.__dict__,
+        )
     return instance
 
 
-def new_parser_from_dataclass(
+def add_dataclass_fields_to_parser(
     dataclass_type: Type[T_DataclassInstance],
+    *,
     parser: Optional[ArgumentParser],
+    list_parsing: ListParsing = "argparse",
 ) -> ArgumentParser:
     if parser is None:
         parser = ArgumentParser()
@@ -79,13 +99,21 @@ def new_parser_from_dataclass(
             msg = f"Invalid field {field.name}: found values for default and default_factory."
             raise ValueError(msg)
 
-        kwds.update(_get_kwds_for_type(field.type))
+        inner_kwds = _get_kwds_for_type(field.type, list_parsing)
+        kwds.update(inner_kwds)
         parser.add_argument(*posargs, **kwds)
 
     return parser
 
 
-def _get_kwds_for_type(field_type: Any) -> Dict[str, Any]:
+@deprecated_alias(add_dataclass_fields_to_parser)
+def new_parser_from_dataclass(*args, **kwargs): ...
+
+
+def _get_kwds_for_type(
+    field_type: Any,
+    list_parsing: ListParsing = "argparse",
+) -> Dict[str, Any]:
     kwds = {}
 
     type_origin = get_origin(field_type)
@@ -101,17 +129,49 @@ def _get_kwds_for_type(field_type: Any) -> Dict[str, Any]:
             msg = f"Invalid argument {field_type=}. (expected homogeneous types in {type_origin})"
             raise TypeError(msg)
 
-    if field_type in _SCALARS_TARGET_TYPES or type_origin in (
-        Literal,
-        Optional,
-        UnionType,
-        Union,
+    if (
+        (field_type in _SCALARS_TARGET_TYPES)
+        or (
+            type_origin
+            in (
+                Literal,
+                Optional,
+                UnionType,
+                Union,
+            )
+        )
+        or (type_origin in (list, Iterable) and list_parsing == "brackets")
     ):
-        kwds.update(_get_kwds_for_scalar_type(field_type, field_type))
-    elif type_origin is list:
+        inner_kwds = _get_kwds_for_scalar_type(field_type, field_type, list_parsing)
+        kwds.update(inner_kwds)
+
+    elif type_origin in (list, Iterable):
         item_type = type_args[0]
-        kwds.update(_get_kwds_for_scalar_type(item_type, field_type))
-        kwds["nargs"] = "*"
+        inner_kwds = _get_kwds_for_scalar_type(item_type, field_type, list_parsing)
+        inner_kwds["nargs"] = "*"
+
+        # TODO: rm
+        # if list_parsing == "argparse":
+        #     kwds["nargs"] = "*"
+        # elif list_parsing == "brackets":
+        #     parse_fn = inner_kwds.pop("type")
+
+        #     def brackets_parse(x: str) -> Any:
+        #         x = (
+        #             x.strip()
+        #             .removeprefix("[")
+        #             .removesuffix("]")
+        #             .removesuffix(",")
+        #             .strip()
+        #         )
+        #         return list(map(parse_fn, x.split(",")))
+
+        #     inner_kwds["type"] = brackets_parse
+        # else:
+        #     msg = f"Invalid argument {list_parsing=}. (expected one of {get_args(ListParsing)})"
+        #     raise ValueError(msg)
+
+        kwds.update(inner_kwds)
     else:
         msg = f"Unsupported type {field_type}."
         raise TypeError(msg)
@@ -119,11 +179,20 @@ def _get_kwds_for_type(field_type: Any) -> Dict[str, Any]:
     return kwds
 
 
-def _get_kwds_for_scalar_type(type_: Any, from_field_type: Any) -> Dict[str, Any]:
+def _get_kwds_for_scalar_type(
+    type_: Any, from_field_type: Any, list_parsing: ListParsing
+) -> Dict[str, Any]:
     type_origin = get_origin(type_)
     kwds = {}
 
-    if type_ in _SCALARS_TARGET_TYPES or type_origin in (UnionType, Union, Optional):
+    if (
+        type_ in _SCALARS_TARGET_TYPES
+        or type_origin in (UnionType, Union, Optional)
+        or (
+            get_origin(from_field_type) in (list, Iterable)
+            and list_parsing == "brackets"
+        )
+    ):
         pass
     elif type_origin is Literal:
         type_args = get_args(type_)
@@ -132,7 +201,7 @@ def _get_kwds_for_scalar_type(type_: Any, from_field_type: Any) -> Dict[str, Any
         msg = f"Unsupported dataclass member type {type_} from {from_field_type}."
         raise TypeError(msg)
 
-    kwds["type"] = parse_to(type_)  # type: ignore
+    kwds["type"] = parse_to(type_, list_parsing=list_parsing)  # type: ignore
     return kwds
 
 
@@ -143,6 +212,7 @@ def parse_to(
     true_values: Union[str, Iterable[str]] = DEFAULT_TRUE_VALUES,
     false_values: Union[str, Iterable[str]] = DEFAULT_FALSE_VALUES,
     none_values: Union[str, Iterable[str]] = DEFAULT_NONE_VALUES,
+    list_parsing: ListParsing = "argparse",
 ) -> Callable[[str], T]:
     """Returns a callable that convert string value to target type safely.
 
@@ -155,6 +225,7 @@ def parse_to(
         true_values=true_values,
         false_values=false_values,
         none_values=none_values,
+        list_parsing=list_parsing,
     )
 
 
@@ -166,6 +237,7 @@ def str_to_type(
     true_values: Union[str, Iterable[str]] = DEFAULT_TRUE_VALUES,
     false_values: Union[str, Iterable[str]] = DEFAULT_FALSE_VALUES,
     none_values: Union[str, Iterable[str]] = DEFAULT_NONE_VALUES,
+    list_parsing: ListParsing = "argparse",
 ) -> T:
     """Convert string values to target type safely. Intended for argparse arguments.
 
@@ -181,6 +253,7 @@ def str_to_type(
         true_values=true_values,
         false_values=false_values,
         none_values=none_values,
+        list_parsing=list_parsing,
     )
     if isinstance(result, Exception):
         raise result
@@ -303,12 +376,14 @@ def _str_to_type_impl(
     true_values: Union[str, Iterable[str]] = DEFAULT_TRUE_VALUES,
     false_values: Union[str, Iterable[str]] = DEFAULT_FALSE_VALUES,
     none_values: Union[str, Iterable[str]] = DEFAULT_NONE_VALUES,
+    list_parsing: ListParsing = "argparse",
 ) -> Union[T, Exception]:
     kwds: Dict[str, Any] = dict(
         case_sensitive=case_sensitive,
         true_values=true_values,
         false_values=false_values,
         none_values=none_values,
+        list_parsing=list_parsing,
     )
     if target_type in _SCALARS_TARGET_TYPES:
         return _str_to_scalar_impl(x, target_type, **kwds)
@@ -329,7 +404,31 @@ def _str_to_type_impl(
             raise ValueError(msg)
         return scalar
 
-    elif getattr(target_type, "__name__", None) == "Optional":
+    if origin in (list, Iterable):
+        if list_parsing != "brackets":
+            raise ValueError
+
+        args = get_args(target_type)
+
+        if len(args) == 0:
+            target_item_type = str
+        elif len(args) == 1:
+            target_item_type = args[0]
+        else:
+            raise ValueError
+
+        x = re.sub(r"^\s*\[\s*(|.*[^,\s])(|\s*,)\s*\]\s*$", r"\1", x)
+        x_list = x.split(",")
+
+        y_list = []
+        for xi in x_list:
+            yi = _str_to_type_impl(xi, target_item_type, **kwds)  # type: ignore
+            if isinstance(yi, Exception):
+                return yi
+            y_list.append(yi)
+        return y_list  # type: ignore
+
+    if getattr(target_type, "__name__", None) == "Optional":
         args = (None,) + get_args(target_type)
     elif origin == Union or origin.__name__ in ("Union", "UnionType"):  # type: ignore
         args = get_args(target_type)
@@ -362,23 +461,31 @@ def _str_to_scalar_impl(
     true_values: Union[str, Iterable[str]] = DEFAULT_TRUE_VALUES,
     false_values: Union[str, Iterable[str]] = DEFAULT_FALSE_VALUES,
     none_values: Union[str, Iterable[str]] = DEFAULT_NONE_VALUES,
+    list_parsing: ListParsing = "argparse",
 ) -> Any:
+    del list_parsing
     if target_type is str:
         return x
+
     elif target_type is int:
         try:
             return int(x)
         except ValueError as err:
             return err
+
     elif target_type is float:
         try:
             return float(x)
         except ValueError as err:
             return err
+
     elif target_type in (None, NoneType):
         return _str_to_none_impl(
-            x, case_sensitive=case_sensitive, none_values=none_values
+            x,
+            case_sensitive=case_sensitive,
+            none_values=none_values,
         )
+
     elif target_type is bool:
         return _str_to_bool_impl(
             x,
@@ -387,7 +494,8 @@ def _str_to_scalar_impl(
             false_values=false_values,
         )
     else:
-        raise ValueError(f"Invalid argument {target_type=}. (unsupported type)")
+        msg = f"Invalid argument {target_type=}. (unsupported type)"
+        raise ValueError(msg)
 
 
 def _str_to_bool_impl(